Fix OverflowException in sync state and improve HTTP handling
- Add validation for negative values in ServerInfo.UpdateFolderSyncState
- Add PruneStaleFolderStates method to clean up stale folder states
- Add diagnostic event for invalid sync data detection
- Fix Connected Devices display showing -1 when no devices configured
- Improve HTTP request handling in SyncthingEventsListener and SyncthingMonitorService using HttpRequestMessage
- Fix MudBlazor 8.x warnings (Dense, Title, IsInitiallyExpanded attributes)
- Fix obsolete WatchSystemPreference warning in App.razor
- Remove unnecessary System.Text.Json package from WebUI
- Fix uninitialized field warnings in Profiling, Statistics, NavMenu

+++ src/CreatioHelper.Contracts/Responses/SyncResponses.cs
@@ -33,8 +33,8 @@ public class SyncFolderDto
     public string State { get; set; } = string.Empty;
     public long GlobalBytes { get; set; }
     public long LocalBytes { get; set; }
-    public int GlobalFiles { get; set; }
-    public int LocalFiles { get; set; }
+    public long GlobalFiles { get; set; }
+    public long LocalFiles { get; set; }
     public DateTime LastScan { get; set; }
     public DateTime LastSync { get; set; }
     public List<string> DeviceIds { get; set; } = new();
